Separate BottomSheet and adapters in separate packages
