+++ app/src/main/java/chat/rocket/android/chatroom/ui/ChatRoomAdapter.kt
@@ -9,6 +9,8 @@ import android.widget.ImageView
 import android.widget.TextView
 import chat.rocket.android.R
 import chat.rocket.android.chatroom.presentation.ChatRoomPresenter
+import chat.rocket.android.chatroom.ui.bottomsheet.adapter.ActionListAdapter
+import chat.rocket.android.chatroom.ui.bottomsheet.BottomSheetMenu
 import chat.rocket.android.chatroom.viewmodel.AttachmentType
 import chat.rocket.android.chatroom.viewmodel.MessageViewModel
 import chat.rocket.android.player.PlayerActivity
